Buying from shop implemented

+++ My project/Assets/Scripts/SO/ItemSO.cs
@@ -11,13 +11,14 @@ public class ItemSO : ScriptableObject
     [SerializeField] private Image _image;
     [SerializeField] private Sprite _sprite;
     [SerializeField] private AnimationClip[] _animations;
-    [SerializeField] private float _price = 0;
+    [SerializeField] private int  _price = 0;
     [SerializeField] private float _stat = 0;
 
     public string GetName{ get { return _name;} }
     public Image GetImage{ get { return _image;  } }
     public Sprite GetSprite { get { return _sprite;  } }
     public AnimationClip[] GetAnimations { get { return _animations;  } }
-    public float GetPrice { get { return _price;  } }
+    public int GetPrice { get { return _price;  } }
     public float GetStat { get { return _stat;  } }
+
 }
